Refactor: Rename SyncMemberFacesToKnowledgeService to SyncMemberFacesToFaceService
- Renamed directory  to .
- Renamed  to .
- Renamed  to .
- Updated class names and namespaces within the renamed files.
- Updated usage in  to reflect the new naming and endpoint.
