+++ apps/backend/src/Application/MemberFaces/Commands/SyncMemberFacesToFaceService/SyncMemberFacesToFaceServiceCommand.cs
@@ -1,0 +1,8 @@
+using backend.Application.Common.Models;
+
+namespace backend.Application.MemberFaces.Commands.SyncMemberFacesToFaceService;
+
+public record SyncMemberFacesToFaceServiceCommand : IRequest<Result<Unit>>
+{
+    public Guid? FamilyId { get; init; }
+}
