--- apps/backend/src/Application/MemberFaces/Commands/SyncMemberFacesToKnowledgeService/SyncMemberFacesToKnowledgeServiceCommand.cs
@@ -1,8 +1,0 @@
-using backend.Application.Common.Models;
-
-namespace backend.Application.MemberFaces.Commands.SyncMemberFacesToKnowledgeService;
-
-public record SyncMemberFacesToKnowledgeServiceCommand : IRequest<Result<Unit>>
-{
-    public Guid? FamilyId { get; init; }
-}
